조회수 기능을 추가합니다.

+++ src/main/java/com/hwans/apiserver/dto/blog/PostDto.java
@@ -51,5 +51,5 @@ public class PostDto implements Serializable {
     @ApiModelProperty(value = "좋아요 수")
     int likeCount;
     @ApiModelProperty(value = "조회수")
-    Long hits;
+    int hits;
 }
